Subset a finger tree by position or element name

Subset a finger tree by position or element name

Usage

# S3 method for class 'FingerTree'
x[i, ...]

Arguments

x

FingerTree.

i

Positive integer indices or character element names.

...

Unused.

Value

A new FingerTree containing selected elements in query order. For character indexing, missing names are represented as `NULL` elements.

Examples

t <- tree_from(letters[1:6])
s <- t[c(2, 4, 6)]
cat_m <- MeasureMonoid(paste0, "", as.character)
reduce_left(s, cat_m)
#> [1] "bdf"

# Empty index returns empty tree
attr(t[integer(0)], "measures")$.size
#> [1] 0

# Character indexing by element names
tn <- tree_from(setNames(as.list(letters[1:4]), c("w", "x", "y", "z")))
out <- tn[c("y", "missing", "w")]
reduce_left(out, MeasureMonoid(paste0, "", function(el) if(is.null(el)) "_" else el))
#> [1] "ca"